Refactor LoRA surgery with Strategy pattern
Extract 8 surgical operations from LoRASurgeon (1,054 lines) into
shared/evolutionary/surgery/ package with Protocol-based strategies:
- 8 operation classes in surgery/operations/
- Decorator-based static registry
- LoRASurgeon reduced to thin orchestrator (232 lines)
- Original lora_surgery.py becomes 53-line backward-compat shim
- All 60 existing tests pass (40 surgery + 20 karpathy integration)
